fix: resolve clippy warnings across the codebase
Addressed 20+ clippy warnings to improve code quality:

- Fixed collapsed if statements in aspen-core and aspen-dns
- Resolved NodeId import shadowing in aspen-api
- Refactored broadcast_blob_announcement to use BlobAnnouncementParams struct (too many args)
- Removed unnecessary .clone() calls on Copy types in aspen-forge
- Removed unused type aliases (ClusterControllerHandle, KeyValueStoreHandle)
- Added NodeComponents type alias to simplify complex return type
- Fixed module inception warning in node/tests.rs
- Prefixed unused variables with underscore
- Applied various auto-fixes for better code style

Remaining warnings require deeper architectural changes and can be
addressed in future refactoring efforts.

+++ crates/aspen-raft/src/storage_shared.rs
@@ -1232,7 +1232,7 @@ impl SharedRedbStorage {
         let actual_lease_id = if lease_id == 0 {
             // Simple ID generation using timestamp + random component
             let now = now_unix_ms();
-            let random_component = (now % 1000000) as u64;
+            let random_component = now % 1000000;
             now * 1000 + random_component
         } else {
             lease_id
